+++ example/lib/shared/decoration/spikes.dart
@@ -1,14 +1,10 @@
-import 'dart:async' as async;
-
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/util/priority_layer.dart';
 import 'package:example/manual_map/dungeon_map.dart';
 import 'package:example/shared/util/common_sprite_sheet.dart';
 
 class Spikes extends GameDecoration with Sensor {
-  async.Timer? timer;
-
-  bool isTick = false;
+  double dt = 0;
 
   Spikes(Vector2 position)
       : super.withSprite(
@@ -16,17 +12,20 @@ class Spikes extends GameDecoration with Sensor {
           position: position,
           width: DungeonMap.tileSize / 1.5,
           height: DungeonMap.tileSize / 1.5,
-        );
+        ) {
+    setupSensorArea(intervalCheck: 500);
+  }
+
+  @override
+  void update(double dt) {
+    this.dt = dt;
+    super.update(dt);
+  }
 
   @override
   void onContact(GameComponent component) {
-    if (timer == null) {
-      if (component is Attackable) {
-        component.receiveDamage(10, 1);
-        timer = async.Timer(Duration(milliseconds: 500), () {
-          timer = null;
-        });
-      }
+    if (component is Attackable) {
+      component.receiveDamage(10, 1);
     }
   }