+++ lib/util/extensions/player/player_extensions.dart
@@ -4,7 +4,6 @@ import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/enemy/enemy.dart';
 import 'package:bonfire/player/player.dart';
 import 'package:bonfire/util/extensions/game_component_extensions.dart';
-import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
 extension PlayerExtensions on Player {