+++ lib/util/mixins/sensor.dart
@@ -6,11 +6,11 @@ import 'package:bonfire/collision/collision_area.dart';
 import 'package:bonfire/collision/collision_config.dart';
 import 'package:bonfire/collision/object_collision.dart';
 import 'package:bonfire/util/vector2rect.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
 import '../interval_tick.dart';
 
-Paint _paintSensor = Paint()..color = Colors.red.withOpacity(0.5);
+Paint _paintSensor = Paint()..color = Color(0xFFF44336).withOpacity(0.5);
 
 /// Mixin responsible for adding trigger to detect other objects above
 mixin Sensor on GameComponent {