+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# [next]
+# 1.11.0
 - Improvements in `Sensor` mixin.
 - Add support to flip vertical, flip horizontal and rotate in Tiled.
 - Update flame to `1.0.0-releasecandidate.17`